+++ source/extensions/filters/network/thrift_proxy/router/router_impl.cc
@@ -25,7 +25,8 @@ RouteEntryImplBase::RouteEntryImplBase(
     : cluster_name_(route.route().cluster()),
       config_headers_(Http::HeaderUtility::buildHeaderDataVector(route.match().headers())),
       rate_limit_policy_(route.route().rate_limits()),
-      strip_service_name_(route.route().strip_service_name()) {
+      strip_service_name_(route.route().strip_service_name()),
+      cluster_header_(route.route().cluster_header()) {
   if (route.route().has_metadata_match()) {
     const auto filter_it = route.route().metadata_match().filter_metadata().find(
         Envoy::Config::MetadataFilters::get().ENVOY_LB);
@@ -211,24 +212,25 @@ FilterStatus Router::messageBegin(MessageMetadataSharedPtr metadata) {
 
   route_entry_ = route_->routeEntry();
 
-  Upstream::ThreadLocalCluster* cluster = cluster_manager_.get(route_entry_->clusterName());
+  const std::string& cluster_name = getClusterName(metadata);
+
+  Upstream::ThreadLocalCluster* cluster = cluster_manager_.get(cluster_name);
   if (!cluster) {
-    ENVOY_STREAM_LOG(debug, "unknown cluster '{}'", *callbacks_, route_entry_->clusterName());
-    callbacks_->sendLocalReply(
-        AppException(AppExceptionType::InternalError,
-                     fmt::format("unknown cluster '{}'", route_entry_->clusterName())),
-        true);
+    ENVOY_STREAM_LOG(debug, "unknown cluster '{}'", *callbacks_, cluster_name);
+    callbacks_->sendLocalReply(AppException(AppExceptionType::InternalError,
+                                            fmt::format("unknown cluster '{}'", cluster_name)),
+                               true);
     return FilterStatus::StopIteration;
   }
 
   cluster_ = cluster->info();
-  ENVOY_STREAM_LOG(debug, "cluster '{}' match for method '{}'", *callbacks_,
-                   route_entry_->clusterName(), metadata->methodName());
+  ENVOY_STREAM_LOG(debug, "cluster '{}' match for method '{}'", *callbacks_, cluster_name,
+                   metadata->methodName());
 
   if (cluster_->maintenanceMode()) {
     callbacks_->sendLocalReply(
         AppException(AppExceptionType::InternalError,
-                     fmt::format("maintenance mode for cluster '{}'", route_entry_->clusterName())),
+                     fmt::format("maintenance mode for cluster '{}'", cluster_name)),
         true);
     return FilterStatus::StopIteration;
   }
@@ -247,11 +249,11 @@ FilterStatus Router::messageBegin(MessageMetadataSharedPtr metadata) {
   ASSERT(protocol != ProtocolType::Auto);
 
   Tcp::ConnectionPool::Instance* conn_pool = cluster_manager_.tcpConnPoolForCluster(
-      route_entry_->clusterName(), Upstream::ResourcePriority::Default, this);
+      cluster_name, Upstream::ResourcePriority::Default, this);
   if (!conn_pool) {
     callbacks_->sendLocalReply(
         AppException(AppExceptionType::InternalError,
-                     fmt::format("no healthy upstream for '{}'", route_entry_->clusterName())),
+                     fmt::format("no healthy upstream for '{}'", cluster_name)),
         true);
     return FilterStatus::StopIteration;
   }
@@ -504,6 +506,24 @@ void Router::UpstreamRequest::onResetStream(Tcp::ConnectionPool::PoolFailureReas
   default:
     NOT_REACHED_GCOVR_EXCL_LINE;
   }
+}
+
+std::string Router::getClusterName(const MessageMetadataSharedPtr& metadata) const {
+  const auto& cluster_name = route_entry_->clusterName();
+  if (!cluster_name.empty()) {
+    return cluster_name;
+  }
+
+  const auto& cluster_header = route_entry_->clusterHeader();
+  if (!cluster_header.get().empty()) {
+    const auto& headers = metadata->headers();
+    const auto* entry = headers.get(cluster_header);
+    if (entry != nullptr) {
+      return std::string(entry->value().getStringView());
+    }
+  }
+
+  return "";
 }
 
 } // namespace Router